Оптимизировал снятие дампа и сделал возможность восстановления
Дамп снимается в папку
Распараллеливание по количеству ядер

+++ db_backup.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-import gzip, os, time, datetime, sys,
-from sh import pg_dump
+
+import gzip, os, time, datetime, sys, argparse
+from sh import pg_dump, pg_restore
 # import telebot
 # from telebot import types, apihelper
 
@@ -10,6 +11,7 @@ db_user = 'postgres'
 db_pass = ""
 disk_need_size = 20
 finish_day = 7
+cpu = 4
 path = "/mnt/backup"
 system_mount_path = '/mnt/backup'
 list_db = ['one', 'two']
@@ -20,27 +22,31 @@ current_time = time.time()
 # bot = telebot.TeleBot('')
 # apihelper.proxy = {'https':'socks5://'}
 
-for i in os.listdir(path):
-    if '.gz' in i:
-        creation_time = os.path.getctime(i)
-        if (current_time - creation_time) // (24 * 3600) > finish_day:
-            os.remove(i)
+parser = argparse.ArgumentParser()
+parser.add_argument('-s', dest='state', default='dump', help="Dump or restore")
+parser.add_argument('-f', dest='files_dump', help="Файл для восстановления")
+parser.add_argument('-d', dest='database', help="База для восстановления")
+args = parser.parse_args()
 
-file_system_mount = os.statvfs(system_mount_path)
-disk_size = int((file_system_mount.f_bsize * file_system_mount.f_bavail) / 10**9)
+if args.state == 'dump':
+    for i in os.listdir(path):
+        if '.gz' in i:
+            creation_time = os.path.getctime(i)
+            if (current_time - creation_time) // (24 * 3600) > finish_day:
+                os.remove(i)
+    file_system_mount = os.statvfs(system_mount_path)
+    disk_size = int((file_system_mount.f_bsize * file_system_mount.f_bavail) / 1000000000)
+    if disk_size <= disk_need_size:
+        message = "На " + db_host + " нехватает места для создания дампа"
+        bot.send_message(chat_id, message)
+        sys.exit()
+    try:
+        for database in list_db:
+            current_datetime = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M")
+            pg_dump('-h', db_host, '-U', db_user, '-d', database, '--encoding=UTF8', '-j', cpu, '--format=directory', '-f', 'backup_' + db_host + '_' + database + '_' + current_datetime)
+    except Exception as e:
+        message = "На " + db_host + " не удалось создать дамп"
+        bot.send_message(chat_id, message)
 
-if disk_size <= disk_need_size:
-    # message = "На " + db_host + " нехватает места для создания дампа"
-    # bot.send_message(chat_id, message)
-    sys.exit()
-
-
-try:
-    for i in list_db:
-        current_datetime = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M")
-        with gzip.open('backup_' + db_host + '_' + i + '_' + current_datetime + '.gz', 'wb') as f:
-            pg_dump('-h', db_host, '-U', db_user, i, _out=f)
-except Exception as e:
-    # message = "На " + db_host + " не удалось создать дамп"
-    # bot.send_message(chat_id, message)
-    sys.exit()
+if args.state == 'restore':
+    pg_restore('-h', db_host, '-U', db_user, '-d', args.database, '-j', cpu, args.files_dump)
